refactor create_processor_job into common module
