+++ common/data_refinery_common/test_job_management.py
@@ -6,11 +6,11 @@ from typing import List
 from unittest.mock import patch, call
 from urllib.error import URLError
 
-from data_refinery_workers.downloaders import utils
+from data_refinery_common.job_management import create_processor_job_for_original_files
 
 
 class UtilsTestCase(TestCase):
-    @tag('downloaders')
+    @tag('processor')
     def test_no_jobs_to_create(self):
         """Make sure this function doesn't raise an exception with no files."""
         utils.create_processor_job_for_original_files([])
