+++ foreman/data_refinery_foreman/foreman/management/commands/create_missing_processor_jobs.py
@@ -22,100 +22,13 @@ from data_refinery_common.models import (
     DownloaderJobOriginalFileAssociation,
     ProcessorJobOriginalFileAssociation
 )
+from data_refinery_common.job_management import create_processor_jobs_for_original_files, create_processor_job_for_original_files
 
 
 logger = get_and_configure_logger(__name__)
 
 BLACKLISTED_EXTENSIONS = ["xml", "chp", "exp"]
 
-
-def delete_if_blacklisted(original_file: OriginalFile) -> OriginalFile:
-    extension = original_file.filename.split(".")[-1]
-    if extension.lower() in BLACKLISTED_EXTENSIONS:
-        logger.debug("Original file had a blacklisted extension of %s, skipping",
-                     extension,
-                     original_file=original_file.id)
-
-        original_file.delete_local_file()
-        original_file.is_downloaded = False
-        original_file.save()
-        return None
-
-    return OriginalFile
-
-
-def create_processor_jobs_for_original_files(original_files: List[OriginalFile],
-                                             volume_index: int):
-    """
-    Create a processor jobs and queue a processor task for samples related to an experiment.
-    """
-    for original_file in original_files:
-        sample_object = original_file.samples.first()
-
-        if not delete_if_blacklisted(original_file):
-            continue
-
-        pipeline_to_apply = determine_processor_pipeline(sample_object, original_file)
-
-        if pipeline_to_apply == ProcessorPipeline.NONE:
-            logger.info("No valid processor pipeline found to apply to sample.",
-                        sample=sample_object.id,
-                        original_file=original_files[0].id)
-            original_file.delete_local_file()
-            original_file.is_downloaded = False
-            original_file.save()
-        else:
-            processor_job = ProcessorJob()
-            processor_job.pipeline_applied = pipeline_to_apply.value
-            processor_job.ram_amount = determine_ram_amount(sample_object, processor_job)
-            processor_job.volume_index = volume_index
-            processor_job.save()
-
-            assoc = ProcessorJobOriginalFileAssociation()
-            assoc.original_file = original_file
-            assoc.processor_job = processor_job
-            assoc.save()
-
-            logger.debug("Queuing processor job.",
-                         processor_job=processor_job.id,
-                         original_file=original_file.id)
-
-            send_job(pipeline_to_apply, processor_job)
-
-
-def create_processor_job_for_original_files(original_files: List[OriginalFile],
-                                            volume_index: int):
-    """
-    Create a processor job and queue a processor task for sample related to an experiment.
-    """
-    # If there's no original files then we've created all the jobs we need to!
-    if len(original_files) == 0:
-        return
-    # For anything that has raw data there should only be one Sample per OriginalFile
-    sample_object = original_files[0].samples.first()
-    pipeline_to_apply = determine_processor_pipeline(sample_object, original_files[0])
-    if pipeline_to_apply == ProcessorPipeline.NONE:
-        logger.info("No valid processor pipeline found to apply to sample.",
-                    sample=sample_object.id,
-                    original_file=original_files[0].id)
-        for original_file in original_files:
-            original_file.delete_local_file()
-            original_file.is_downloaded = False
-            original_file.save()
-    else:
-        processor_job = ProcessorJob()
-        processor_job.pipeline_applied = pipeline_to_apply.value
-        processor_job.ram_amount = determine_ram_amount(sample_object, processor_job)
-        processor_job.volume_index = volume_index
-        processor_job.save()
-        for original_file in original_files:
-            assoc = ProcessorJobOriginalFileAssociation()
-            assoc.original_file = original_file
-            assoc.processor_job = processor_job
-            assoc.save()
-        logger.debug("Queuing processor job.",
-                     processor_job=processor_job.id)
-        send_job(pipeline_to_apply, processor_job)
 
 
 def find_volume_index_for_dl_job(job: DownloaderJob) -> int:
