+++ foreman/data_refinery_foreman/foreman/management/commands/rerun_salmon_old_samples.py
@@ -29,54 +29,9 @@ from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.rna_seq import get_quant_results_for_experiment
 from data_refinery_common.utils import get_env_variable, get_active_volumes
+from data_refinery_common.job_management import create_processor_job_for_original_files
 
 logger = get_and_configure_logger(__name__)
-
-from data_refinery_common.job_lookup import determine_processor_pipeline, determine_ram_amount
-
-def create_processor_job_for_original_files(original_files: List[OriginalFile],
-                                            downloader_job: DownloaderJob=None):
-    """
-    Create a processor job and queue a processor task for sample related to an experiment.
-    """
-
-    # If there's no original files then we've created all the jobs we need to!
-    if len(original_files) == 0:
-        return
-
-    # For anything that has raw data there should only be one Sample per OriginalFile
-    sample_object = original_files[0].samples.first()
-    pipeline_to_apply = determine_processor_pipeline(sample_object, original_files[0])
-
-    if pipeline_to_apply == ProcessorPipeline.NONE:
-        logger.info("No valid processor pipeline found to apply to sample.",
-                    sample=sample_object.id,
-                    original_file=original_files[0].id)
-        for original_file in original_files:
-            original_file.delete_local_file()
-            original_file.is_downloaded = False
-            original_file.save()
-    else:
-        processor_job = ProcessorJob()
-        processor_job.pipeline_applied = pipeline_to_apply.value
-        processor_job.ram_amount = determine_ram_amount(sample_object, processor_job)
-        processor_job.save()
-        for original_file in original_files:
-            assoc = ProcessorJobOriginalFileAssociation()
-            assoc.original_file = original_file
-            assoc.processor_job = processor_job
-            assoc.save()
-
-        logger.debug("Queuing processor job.",
-                     processor_job=processor_job.id)
-
-        try:
-            send_job(pipeline_to_apply, processor_job)
-        except:
-            # If we cannot queue the job now the Foreman will do
-            # it later.
-            pass
-
 
 def update_salmon_versions(experiment: Experiment):
     quant_results = get_quant_results_for_experiment(experiment)
